Moved to pulp-common - shared by CLI and agent.
