+++ common/pulp/common/lock.py
@@ -1,15 +1,3 @@
-#
-# Copyright (c) 2011 Red Hat, Inc.
-#
-# This software is licensed to you under the GNU General Public
-# License as published by the Free Software Foundation; either version
-# 2 of the License (GPLv2) or (at your option) any later version.
-# There is NO WARRANTY for this software, express or implied,
-# including the implied warranties of MERCHANTABILITY,
-# NON-INFRINGEMENT, or FITNESS FOR A PARTICULAR PURPOSE. You should
-# have received a copy of GPLv2 along with this software; if not, see
-# http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt.
-#
 
 """
 Contains locking classes.
@@ -17,11 +5,13 @@ Contains locking classes.
 
 import os
 import fcntl
+
 from threading import RLock
 
 
 class LockFailed(Exception):
     pass
+
 
 class NotLocked(Exception):
     pass
